remove jemalloc prefix on Mac OS builds
This commit removes the jemalloc prefix for Mac OS builds.

On linux, jemalloc by default does not include a prefix. On Mac, by default, it configures itself
to use a `je_` prefix. Some codebases do not expect this prefix. Looking at the homebrew
formula, the default is to explicitly remove the prefix:

https://github.com/Homebrew/homebrew-core/blob/c8f2b60a51be6b4ec039100df3f55d678548f9ac/Formula/jemalloc.rb